+++ ConsoleExample/Models/TestParams.cs
@@ -2,7 +2,7 @@
 
 namespace ConsoleExample.Models
 {
-    public class TestParams : CustomParameters
+    public class TestParams 
     {
         public TestParams()
         {
@@ -11,14 +11,14 @@ namespace ConsoleExample.Models
         public DateTime? Data { get; set; }
         public string DataOne { get; set; }
         public string DataTwo { get; set; }
-        public override void ClearData()
+        public void ClearData()
         {
             DataOne = "";
             DataTwo = "";
             Data = null;
         }
 
-        public override void InitData()
+        public void InitData()
         {
             Data = DateTime.Now;