+++ PRTelegramBot/Extensions/StepService.cs
@@ -1,28 +1,29 @@
 ﻿using Telegram.Bot.Types;
 using PRTelegramBot.Models;
 using System.Collections.Concurrent;
+using PRTelegramBot.Interface;
 
 namespace PRTelegramBot.Extensions
 {
     /// <summary>
     /// Позволяет пользователю выполнять команды пошагово
     /// </summary>
-    public static class Step
+    public static class StepService
     {
         /// <summary>
         /// Список шагов для пользователя
         /// </summary>
-        static ConcurrentDictionary<long, StepTelegram> _step = new();
+        static ConcurrentDictionary<long, IExecuteStep> _step = new();
 
         /// <summary>
         /// Регистрация следующего шага
         /// </summary>
         /// <param name="update">Обновление полученное с телеграма</param>
         /// <param name="command">Следующая команда которая должна быть выполнена</param>
-        public static void RegisterNextStep(this Update update, StepTelegram command)
+        public static void RegisterStepHandler(this Update update, IExecuteStep command)
         {
             long userId = update.GetChatId();
-            update.ClearStepUser();
+            update.ClearStepUserHandler();
             _step.AddOrUpdate(userId, command, (_, existingData) => command);
         }
 
@@ -31,7 +32,7 @@ namespace PRTelegramBot.Extensions
         /// </summary>
         /// <param name="update">Обновление полученное с телеграма</param>
         /// <returns>шаг или null</returns>
-        public static StepTelegram? GetStepOrNull(this Update update)
+        public static IExecuteStep? GetStepHandler(this Update update)
         {
             long userId = update.GetChatId();
             return (_step.TryGetValue(userId, out var data)) ? data : null;
@@ -41,14 +42,13 @@ namespace PRTelegramBot.Extensions
         /// Очищает шаги пользователя
         /// </summary>
         /// <param name="update">Обновление полученное с телеграма</param>
-        public static void ClearStepUser(this Update update)
+        public static void ClearStepUserHandler(this Update update)
         {
             long userId = update.GetChatId();
-            if (update.HasStep())
-            {
-                StepTelegram deleteData;
-                _step.Remove(userId,out deleteData);
-            }
+
+            if (update.HasStepHandler())
+                _step.Remove(userId, out _);
+            
         }
 
         /// <summary>
@@ -56,24 +56,10 @@ namespace PRTelegramBot.Extensions
         /// </summary>
         /// <param name="update">Обновление полученное с телеграма</param>
         /// <returns>True/false</returns>
-        public static bool HasStep(this Update update)
+        public static bool HasStepHandler(this Update update)
         {
             long userId = update.GetChatId();
-            if (!_step.ContainsKey(userId))
-                return false;
-
-            var data = update.GetStepOrNull();
-            if (data.ExpiriedTime == null)
-                return true;
-
-            if (DateTime.Now > data.ExpiriedTime)
-            {
-                data.ExpiriedTime = null;
-                StepTelegram deleteData;
-                _step.Remove(userId, out deleteData);
-                return false;
-            }
-            return true;
+            return _step.ContainsKey(userId);
         }
     }