+++ PRTelegramBot/Models/Enums/ResultExecuteStep.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PRTelegramBot.Models.Enums
+{
+    public enum ResultExecuteStep
+    {
+        None = 1,
+        Success,
+        Failure,
+        ExpiredTime
+    }
+}